feat(store): split MLP store into slices
